Amendments to work with fifo

+++ 1HWH/task_aux/program2.c
@@ -18,34 +18,34 @@ int main(char argc, char *argv[])
         return 0;
     }
 
+    char *fifo_file = argv[1];
+
+    int fd = 0;
+    if ((fd = open(fifo_file, O_RDONLY)) < 0)
+    {
+        printf("Can't open output stream\n");
+        exit(-1);
+    }
+
     pid_t process = fork();
 
     if (process < 0)
     {
-        printf("Fork can't create\n");
+        printf("Can't create fork process\n");
         exit(-1);
     }
     else if (process == 0)
     {
+        if (close(fd) < 0)
+        {
+            printf("Can't close output stream\n");
+            exit(-1);
+        }
+
         char delcom[LEN_PATH] = "rm -rf ";
         strcat(delcom, argv[1]);
 
-        printf("%s\n", delcom);
-
         if ((execl("/bin/bash", "/bin/bash", "-c", delcom) < 0))
-        {
-            printf("Unsuccessful execution of the program\n");
-            exit(-1);
-        }
-    }
-    else
-    {
-        char command[LEN_PATH] = "cat ";
-        strcat(command, argv[1]);
-
-        printf("%s\n", command);
-
-        if ((execl("/bin/bash", "/bin/bash", "-c", command) < 0))
         {
             printf("Unsuccessful execution of the program\n");
             exit(-1);
